backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
cryptography>=42.0.0
supabase==2.9.1
stripe==10.12.0
PyJWT==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
